+++ src/realm/object-store/impl/collection_notifier.cpp
@@ -22,7 +22,9 @@
 #include <realm/object-store/shared_realm.hpp>
 
 #include <realm/db.hpp>
+#include <realm/dictionary.hpp>
 #include <realm/list.hpp>
+#include <realm/set.hpp>
 
 using namespace realm;
 using namespace realm::_impl;
@@ -32,8 +34,8 @@ bool CollectionNotifier::any_related_table_was_modified(TransactionChangeInfo co
     // Check if any of the tables accessible from the root table were
     // actually modified. This can be false if there were only insertions, or
     // deletions which were not linked to by any row in the linking table
-    auto table_modified = [&](auto& tbl) {
-        auto it = info.tables.find(tbl.table_key.value);
+    auto table_modified = [&](auto& outgoing_linkset) {
+        auto it = info.tables.find(outgoing_linkset.table_key.value);
         return it != info.tables.end() && !it->second.modifications_empty();
     };
     return any_of(begin(m_related_tables), end(m_related_tables), table_modified);
@@ -52,7 +54,7 @@ CollectionNotifier::get_modification_checker(TransactionChangeInfo const& info, 
     }
 
     if (m_related_tables.size() == 1) {
-        auto& object_set = info.tables.find(m_related_tables[0].table_key.value)->second;
+        auto& object_set = info.tables.find(m_related_tables.begin()->table_key.value)->second;
         return [&](ObjectChangeSet::ObjectKeyType object_key) {
             return object_set.modifications_contains(object_key);
         };
@@ -61,32 +63,78 @@ CollectionNotifier::get_modification_checker(TransactionChangeInfo const& info, 
     return DeepChangeChecker(info, *root_table, m_related_tables);
 }
 
-void DeepChangeChecker::find_related_tables(std::vector<RelatedTable>& out, Table const& table)
+void DeepChangeChecker::find_related_tables(DeepChangeChecker::RelatedTables& out, Table const& table)
 {
-    auto table_key = table.get_key();
-    if (any_of(begin(out), end(out), [=](auto& tbl) {
-            return tbl.table_key == table_key;
-        }))
-        return;
+    struct LinkInfo {
+        std::vector<ColKey> link_columns;
+        std::vector<TableKey> connected_tables;
+    };
 
-    // We need to add this table to `out` before recurring so that the check
-    // above works, but we can't store a pointer to the thing being populated
-    // because the recursive calls may resize `out`, so instead look it up by
-    // index every time
-    size_t out_index = out.size();
-    out.push_back({table_key, {}});
+    // Build up the complete forward mapping from the back links.
+    // Following forward link columns does not account for TypedLink
+    // values as part of Dictionary<String, Mixed> for example. But
+    // we do not want to assume that all Mixed columns contain links,
+    // so we rely on the fact that if there are any TypedLinks from a
+    // Mixed value, there will be a corresponding backlink column
+    // created at the destination table.
+    std::unordered_map<TableKey, LinkInfo> complete_mapping;
+    Group* group = table.get_parent_group();
+    REALM_ASSERT(group);
+    auto all_table_keys = group->get_table_keys();
+    for (auto key : all_table_keys) {
+        auto cur_table = group->get_table(key);
+        REALM_ASSERT(cur_table);
+        auto incoming_link_columns = cur_table->get_incoming_link_columns();
+        for (auto& incoming_link : incoming_link_columns) {
+            REALM_ASSERT(incoming_link.first);
+            auto& links = complete_mapping[incoming_link.first];
+            links.link_columns.push_back(incoming_link.second);
+            links.connected_tables.push_back(cur_table->get_key());
+        }
+    }
 
-    for (auto col_key : table.get_column_keys()) {
-        auto type = table.get_column_type(col_key);
-        if (type == type_Link || type == type_LinkList) {
-            out[out_index].links.push_back({col_key.value, type == type_LinkList});
-            find_related_tables(out, *table.get_link_target(col_key));
+    // Remove duplicates:
+    // duplicates in link_columns can occur when a Mixed(TypedLink) contain links to different tables
+    // duplicates in connected_tables can occur when there are different link paths to the same table
+    for (auto& kv_pair : complete_mapping) {
+        auto& cols = kv_pair.second.link_columns;
+        std::sort(cols.begin(), cols.end());
+        cols.erase(std::unique(cols.begin(), cols.end()), cols.end());
+        auto& tables = kv_pair.second.connected_tables;
+        std::sort(tables.begin(), tables.end());
+        tables.erase(std::unique(tables.begin(), tables.end()), tables.end());
+    }
+
+    std::vector<TableKey> tables_to_check = {table.get_key()};
+    auto get_out_relationships_for = [&out](TableKey key) -> std::vector<ColKey>& {
+        auto it = find_if(begin(out), end(out), [&](auto&& tbl) {
+            return tbl.table_key == key;
+        });
+        if (it == out.end()) {
+            it = out.insert(out.end(), {key, {}});
+        }
+        return it->links;
+    };
+
+    while (tables_to_check.size()) {
+        auto table_key_to_check = *tables_to_check.begin();
+        tables_to_check.erase(tables_to_check.begin());
+        auto outgoing_links = complete_mapping[table_key_to_check];
+        auto& out_relations = get_out_relationships_for(table_key_to_check);
+        auto& link_columns = complete_mapping[table_key_to_check].link_columns;
+        out_relations.insert(out_relations.end(), link_columns.begin(), link_columns.end());
+        for (auto linked_table_key : outgoing_links.connected_tables) {
+            if (std::find_if(begin(out), end(out), [&](auto&& relation) {
+                    return relation.table_key == linked_table_key;
+                }) == out.end()) {
+                tables_to_check.push_back(linked_table_key);
+            }
         }
     }
 }
 
 DeepChangeChecker::DeepChangeChecker(TransactionChangeInfo const& info, Table const& root_table,
-                                     std::vector<RelatedTable> const& related_tables)
+                                     DeepChangeChecker::RelatedTables const& related_tables)
     : m_info(info)
     , m_root_table(root_table)
     , m_root_table_key(root_table.get_key().value)
@@ -98,11 +146,69 @@ DeepChangeChecker::DeepChangeChecker(TransactionChangeInfo const& info, Table co
 {
 }
 
-bool DeepChangeChecker::check_outgoing_links(TableKey table_key, Table const& table, int64_t obj_key, size_t depth)
+bool DeepChangeChecker::do_check_for_collection_modifications(std::unique_ptr<CollectionBase> coll, size_t depth)
+{
+    REALM_ASSERT(coll);
+    if (auto lst = dynamic_cast<LnkLst*>(coll.get())) {
+        TableRef target = lst->get_target_table();
+        return std::any_of(lst->begin(), lst->end(), [&, this](auto key) {
+            return this->check_row(*target, key.value, depth + 1);
+        });
+    }
+    else if (auto dict = dynamic_cast<Dictionary*>(coll.get())) {
+        auto target = dict->get_target_table();
+        TableRef cached_linked_table;
+        return std::any_of(dict->begin(), dict->end(), [&, this](auto key_value_pair) {
+            Mixed value = key_value_pair.second;
+            if (value.is_type(type_TypedLink)) {
+                auto link = value.get_link();
+                REALM_ASSERT(link);
+                if (!cached_linked_table || cached_linked_table->get_key() != link.get_table_key()) {
+                    cached_linked_table = dict->get_table()->get_parent_group()->get_table(link.get_table_key());
+                    REALM_ASSERT_EX(cached_linked_table, link.get_table_key().value);
+                }
+                return this->check_row(*cached_linked_table, link.get_obj_key().value, depth + 1);
+            }
+            else if (value.is_type(type_Link)) {
+                REALM_ASSERT(target);
+                return this->check_row(*target, value.get<ObjKey>().value, depth + 1);
+            }
+            return false;
+        });
+    }
+    else if (auto set = dynamic_cast<LnkSet*>(coll.get())) {
+        auto target = set->get_target_table();
+        REALM_ASSERT(target);
+        return std::any_of(set->begin(), set->end(), [&, this](auto obj_key) {
+            return obj_key && this->check_row(*target, obj_key.value);
+        });
+    }
+    else if (auto set = dynamic_cast<Set<Mixed>*>(coll.get())) {
+        TableRef cached_linked_table;
+        return std::any_of(set->begin(), set->end(), [&, this](auto value) {
+            if (value.is_type(type_TypedLink)) {
+                auto link = value.get_link();
+                REALM_ASSERT(link);
+                if (!cached_linked_table || cached_linked_table->get_key() != link.get_table_key()) {
+                    cached_linked_table = set->get_table()->get_parent_group()->get_table(link.get_table_key());
+                    REALM_ASSERT_EX(cached_linked_table, link.get_table_key().value);
+                }
+                return this->check_row(*cached_linked_table, link.get_obj_key().value, depth + 1);
+            }
+            return false;
+        });
+    }
+    // at this point, we have not handled all datatypes
+    REALM_UNREACHABLE();
+    return false;
+}
+
+bool DeepChangeChecker::check_outgoing_links(TableKey table_key, Table const& table, ObjKey obj_key, size_t depth)
 {
     auto it = find_if(begin(m_related_tables), end(m_related_tables), [&](auto&& tbl) {
         return tbl.table_key == table_key;
     });
+
     if (it == m_related_tables.end())
         return false;
     if (it->links.empty())
@@ -110,7 +216,7 @@ bool DeepChangeChecker::check_outgoing_links(TableKey table_key, Table const& ta
 
     // Check if we're already checking if the destination of the link is
     // modified, and if not add it to the stack
-    auto already_checking = [&](int64_t col) {
+    auto already_checking = [&](ColKey col) {
         auto end = m_current_path.begin() + depth;
         auto match = std::find_if(m_current_path.begin(), end, [&](auto& p) {
             return p.obj_key == obj_key && p.col_key == col;
@@ -125,21 +231,18 @@ bool DeepChangeChecker::check_outgoing_links(TableKey table_key, Table const& ta
     };
 
     const Obj obj = table.get_object(ObjKey(obj_key));
-    auto linked_object_changed = [&](OutgoingLink const& link) {
-        if (already_checking(link.col_key))
+    auto linked_object_changed = [&](ColKey const& outgoing_link_column) {
+        if (already_checking(outgoing_link_column))
             return false;
-        if (!link.is_list) {
-            if (obj.is_null(ColKey(link.col_key)))
+        if (!outgoing_link_column.is_collection()) {
+            if (obj.is_null(outgoing_link_column))
                 return false;
-            auto dst = obj.get<ObjKey>(ColKey(link.col_key)).value;
-            return check_row(*table.get_link_target(ColKey(link.col_key)), dst, depth + 1);
+            ObjKey dst = obj.get<ObjKey>(outgoing_link_column);
+            REALM_ASSERT(dst);
+            return check_row(*table.get_link_target(outgoing_link_column), dst.value, depth + 1);
         }
-
-        auto& target = *table.get_link_target(ColKey(link.col_key));
-        auto lvr = obj.get_linklist(ColKey(link.col_key));
-        return std::any_of(lvr.begin(), lvr.end(), [&, this](auto key) {
-            return this->check_row(target, key.value, depth + 1);
-        });
+        auto collection_ptr = obj.get_collection_ptr(outgoing_link_column);
+        return do_check_for_collection_modifications(std::move(collection_ptr), depth);
     };
 
     return std::any_of(begin(it->links), end(it->links), linked_object_changed);
@@ -167,7 +270,7 @@ bool DeepChangeChecker::check_row(Table const& table, ObjKeyType key, size_t dep
     if (it != not_modified.end())
         return false;
 
-    bool ret = check_outgoing_links(table_key, table, key, depth);
+    bool ret = check_outgoing_links(table_key, table, ObjKey(key), depth);
     if (!ret && (depth == 0 || !m_current_path[depth - 1].depth_exceeded))
         not_modified.insert(key);
     return ret;