+++ test/test_lang_bind_helper.cpp
@@ -2000,6 +2000,10 @@ public:
     {
         return false;
     }
+    bool typed_link_change(ColKey, TableKey)
+    {
+        return true;
+    }
 };
 
 struct AdvanceReadTransact {